refactor: Pull itinerary results section out as a LiveComponent
